+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Settings page for database connection configuration. Fixes UILDP-14.
 * Protect settings pages with permissions. Fixes UILDP-30.
 * Factor out common code in `load*` functions. Fixes UILDP-21.
+* Add settings page for saved-query GitHub repo details. Fixes UILDP-31.
 
 ## [1.5.0](https://github.com/folio-org/ui-ldp/tree/v1.5.0) (2021-10-01)